refactor: move RejectsUnknownKeys into a shared write base module
It sat in `write/serializers.py` next to the chatbot settings serializers,
but it is not specific to them -- `apps/api/schema.py` already reaches in for
it, and every future write surface needs the same rule. Move it to a
`write/base.py` that holds no serializers of its own.
